+++ slack-market/commands.rb
@@ -1,2 +1,3 @@
 require 'slack-market/commands/help'
 require 'slack-market/commands/quote'
+require 'slack-market/commands/set'