+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 2/8/2016: [#3](https://github.com/dblock/slack-market/issues/3): Added `set dollars on|off`, respond to $MSFT but not MSFT - [@dblock](https://github.com/dblock).
 * 2/5/2016: Respond to single character stocks with a $ sign, eg. `$F` - [@dblock](https://github.com/dblock).
 * 2/5/2016: Don't respond to channel mentions that contain what looks like a ticker - [@dblock](https://github.com/dblock).
 * 2/4/2016: Initial public release - [@dblock](https://github.com/dblock).